backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
anthropic==0.28.0
resend==2.0.0
gspread==6.1.2
google-auth==2.29.0
httpx==0.27.0
